Update README.md
In the scenario where a developer may land on this repository first, we should consider adding next step guidance after they build and run this code. To get the most value from this repo, as a last step the developer should build and run the test Google iOS app to simulate the App Flip integration.